requirements.txt
attrs==22.2.0
base58==2.1.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
exceptiongroup==1.1.0
idna==3.4
iniconfig==2.0.0
morphys==1.0
msgpack==1.0.4
packaging==23.0
pinata-python==1.0.0
pinatapy-vourhey==0.1.8
pluggy==1.0.0
py-algorand-sdk==2.0.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multiformats-cid==0.4.4
py-multihash==2.0.1
pycparser==2.21
pycryptodomex==3.17
PyNaCl==1.5.0
pytest==7.2.1
python-baseconv==1.2.2
requests==2.28.2
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
varint==1.0.2
